finished writing the tests for all the bitfields
The Bitfield ids are all tested, and if there are any changes they will
have to reflect also in test code, or else it will fail.

+++ fecp/src/main/java/com/ifit/sparky/fecp/interpreter/bitField/converter/BitfieldDataConverter.java
@@ -109,11 +109,11 @@ public abstract class BitfieldDataConverter {
         }
         else if(this.mDataSize == 4)
         {
-            if(data > Long.MAX_VALUE)
+            if(data > Integer.MAX_VALUE)
             {
-                throw new InvalidBitFieldException(data, Long.TYPE);
+                throw new InvalidBitFieldException(data, Integer.TYPE);
             }
-            tempBuff.putLong((long)data);
+            tempBuff.putInt((int)data);
         }
 
         return tempBuff;
